alteracoes de uns dias ai

+++ src/main/java/br/csi/oportunidades/repository/InstituicaoRepository.java
@@ -3,5 +3,5 @@ package br.csi.oportunidades.repository;
 import br.csi.oportunidades.model.Instituicao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InstituicaoRepository extends JpaRepository<Instituicao, Integer> {
+public interface InstituicaoRepository extends JpaRepository<Instituicao, Long> {
 }
